==> script5.py <==
# requires pandas
import pandas

# define file paths
ASequences_file = "/Users/santiagochang/Desktop/JSB/WBcel235/ASequences.tsv"
uORFs_file = "/Users/santiagochang/Desktop/JSB/WBcel235/output_uORF.tsv"
output_file = "/Users/santiagochang/Desktop/JSB/WBcel235/updated_ASequences.tsv"

# input: filtered data frame generated by script4.py
# output: dictionary mapping gene ID to variables
def construct_gene_dictionary(uORFs):
    gene_dictionary = {}
    for _, row in uORFs.iterrows():
        key = row.iloc[0]
        value = row.iloc[1:].tolist()
        gene_dictionary[key] = value
    return gene_dictionary

# input: ASequences data frame, gene dictionary
# output: updated data frame with cDNA sequence, distance of uORF from CDS, kozak sequence
def create_data_frame(ASequences, gene_dictionary):
    new_data = []
    progress = 0
    for i in ASequences.index:
        # print progress in 1% intervals
        if int(i/(len(ASequences) - 1) * 100) == progress:
            print(f'Construction of data frame is {progress}% complete.')
            progress += 1
        gene = ASequences.at[i, 'gene']
        start = ASequences.at[i, 'A_start']
        cdna = (gene_dictionary[gene])[1].upper()
        cds_start = (gene_dictionary[gene])[10]
        distance = int(cds_start) - int(start)
        kozak = cdna[max(start-4, 0):start+3]
        new_data.append([cdna, distance, kozak])
    new_data_frame = pandas.DataFrame(new_data, columns=['cdna', 'distance', 'kozak'])
    ASequences = ASequences.reset_index(drop=True)
    new_data_frame = new_data_frame.reset_index(drop=True)
    updated_data_frame = pandas.concat([ASequences, new_data_frame], axis=1)
    return updated_data_frame

# read in input files
ASequences_df = pandas.read_csv(ASequences_file, sep='\t')
uORFs_df = pandas.read_csv(uORFs_file, sep='\t')
# construct gene dictionary
gene_dict = construct_gene_dictionary(uORFs_df)
# construct updated data frame and output
updated_ASequences = create_data_frame(ASequences_df, gene_dict)
updated_ASequences.to_csv(output_file, sep = '\t', index = False)
